requirements.txt
Flask
flask-cors
easyocr
opencv-python-headless
numpy
Pillow